replace go-ipld-prime-proto with go-codec-dagpb
We're transitioning to the latter codec for dag-pb on ipld-prime, so
go-car should use it too to avoid registering two codecs to handle the
same multicodec code.

Besides swapping the library, we also update go-ipld-prime to grab its
raw codec, to be used for raw nodes.

We also roll our own little version of go-ipld-prime-proto's
AddDagPBSupportToChooser. With go-codec-dagpb, raw nodes don't need a
special prototype, so we need just one special case for PBNode. It also
doesn't seem worthwhile to add API to go-codec-dagpb for what is
essentially three lines of fairly simple code. We can always revisit
that if more downstream users require it too.