Define CAR v2 as a separate go module
Since we are releasing a new major version, define a new go module for
the CAR v2 implementation, following the recommendations here:
- https://blog.golang.org/v2-go-modules

The module is versioned as go 1.15 to maintain compatibility as pointed
out by comment here:
- https://github.com/ipld/go-car/pull/75#discussion_r644634850

Note, in this implementation we are depending on all libraries CAR v1
depended on. Future improvements will consider integration with
ipld/go-ipld-prime primitives.