Implement CAR v2 writer that uses `format.NodeGetter`
Implement a CAR v2 writer, that produces binary structure corresponding
to the specification of car V2, consisting of:
 1. Version prefix
 2. CAR v2 header
 3. Dump of Car v1
 4. Carbs index

The implementation also facilities optional padding before dump of CAR
v1 and Carbs index to provide scope for future optimisations. The
padding is defined as a dedicated type that writes zero-valued bytes for
a given padding size. The CAR v2 header, then captures the offsets from
the beginning of the file for both the CAR v1 dump and Carbs index. This
allows the user to quickly skip to the part of the CAR v2 they need as
well as offset alignment, if necessary, by altering the value of
padding.

Note, this is an intermediate implementation, and does not correctly
count the number of bytes written by the writer, pending the
transfer/refactor of Carbs implementation in this repo. In the meantime,
This implementation depends on Carbs as a go module. The implementation
of extensive writer tests is postponed until Carbs is transferred and
the written byte-count can be returned from the index marshaller.