Implement read-only random access blockstore
Refactor the carbon and carbs packages into a read-only blockstore.

Move the packages to `internal` as needed and remove duplicate reader
implementations.

Improve serialization efficiency for car v2 primitives.

Note the readonly blockstore will be altered in the coming PRs to
understand car v2 format and work transparently. For now we want to
push the refactoring and changes to unblock other parallel workstreams.