Implement CAR introspector API
Implement an API that allows a car file to be introspected, which would
worth for both CAR v1 and v2. When the given reader is a v1, the
`HasIndex` will always be false, and the car v1 size is set to the total
readable bytes in the given reader. When the given reader is a v2, then
the bytes are parsed and corresponding values are fetched from v2
header. Regardless of car version, `Roots` filed is populated, either
from the v1 header, or the v1 dump in car v2.

Add `HasIndex` placeholder in v2 header characteristics. Note the
placeholder always returns true, since the current writer always writes
the index. When the writer is made configurable we then change the
placeholder to do its thing.

Add lazy loading of roots in CAR v2 reader, used by the introspector.

Add utility io function to size the number of readable bytes in a
reader.