Rename `ReadPragma` to `ReadVersion` and drop returning roots
Change the `ReadPragma` signature to only return the version and drop
the roots that _may_ exist if the version is 1. Because, this results in
a less error-prone API, and right now it is unclear if we want
functionality that should return roots. So, we lean towards providing
less functionality unless requested.