consistently use CID hashes for uniqueness
The blockstore used comparisons on the entire CID for Has/Get to succeed,
and the indexes generally used the inner multihash's digest to uniquely
identify indexed CID-block pairs.

The blockstore interface is generally understood to understand CID
uniqueness by hash, and it's what go-ipfs's datastore does, so let's do
the same.

Add a test case for what broke with that inconsistency.

Updates #110.