add carv2.WrapV1 based on io interfaces
This requires index.Generate to work on io.ReadSeeker, which is a good
idea anyway, as it just needs to read and seek. The main advantage of
ReaderAt is that it allows concurrent reads, but index generation is
sequential by nature. Plus, it only needs seeking to skip blocks; it
doesn't need random access to the carv1.

Also make the WrapV1 example test that the resulting carv2 works,
including the index.

Fixes #129.