Remove resumption option for automatic resumption when file exists
When file exists, attempt resumption by default. This seems like a more
user-friendly API and less requirements to explicitly tune knobs.

Explicitly fail of the file is finalized. We technically can append to a
finalized file but that involves changes in multiple places in reader
abstractions. Left a TODO as a feature for future when asked for.

Close off file if construction of ReadWrite blockstore fails after file
is opened. If left open, it causes test failures when t cleans up temp
files.