Remove redundant seek before read or generate index
Assume the read seeker passed in is seeked to the beginning of CAR
payload. This is both for consistency of API and avoid unnecessary
seeks when the reader may already be at the right place.

Address review comments

* Use the file already open to get the stats for resumption purposes. On
getting the stats, because the open flag contains `os.O_CREATE` check
the size of the file as a way to determine wheter we should attempt
resumption or not. In practice this is the same as the code that
explicitly differentiates from non-existing files, since file existence
doesn't matter here; the key thing is the file content. Plus this also
avoids unnecessary errors where the files exists but is empty.

* Add TODOs in places to consider enabling deduplication by default and
optimise computational complexity of roots check. Note, explicitly
enable deduplication by default in a separate PR so that the change is
clearly communicated to dependant clients in case it causes any
unintended side-effects.

* Rename `Equals` to `Matches` to avoid confusion about what it does.

Relates to:
- https://github.com/ipld/go-car/pull/147#discussion_r668587909
- https://github.com/ipld/go-car/pull/147#discussion_r668603050
- https://github.com/ipld/go-car/pull/147#discussion_r668609927