Support resumption in `ReadWrite` blockstore for finalized files
Support resumption from a complete CARv2 file, effectively implementing
append block to a CARv2 file. Note, on resumption the entire file is
re-indexed. So it is undesirable to do this for complete cars.

Fix a bug where `AllKeysChan` failed on files that contained the index
and resumed from. Fix is done by truncating the CARv2 file, removing the
index if it is present when blockstore is constructed when resuming.
This is because if `AllKeysChan` is called on a file that contains the
index, the ReadOnly backing will continue to read until EOF which will
result in error. Since file is re-indexed on resumption and we require
finalize to be called we truncate the file to remove index.

Write header on every put so that resuming from finalized files that put
more blocks without then finalizing works as expected. Otherwise,
because the last put did not finalize and header is not updated any
blocks put will effectively be lost.

Add tests that assert resume with and without finalization. Add tests
that assert read operations work on resumed read-write blockstore.

Fixes: https://github.com/ipld/go-car/issues/155