v2: stop using a symlink for LICENSE.md
Go modules has to support many OSs and filesystems,
so it completely ignores symlinks when creating module zips.
And parts of the Go ecosystem, like pkg.go.dev, consume those zips.

So this symlink actually makes the module technically non-OSS,
since it completely lacks a LICENSE file when downloaded.
pkg.go.dev thus refuses to show its documentation.

One option would be to rename the root LICENSE.md file to just LICENSE,
since cmd/go treats that file in a special way,
including it in subdirectory-module zips.
Unfortunately, we do want the ".md" extension,
since the file is formatted to outline the double license.

As such, simply copy the file. It should be fine as it's mostly static.