avoid allocating on every byte read
We need a ByteReader for some APIs, such as reading CIDs.
However, our entrypoints often don't, such as Reader or ReaderAt.
Thus, we have a type that does the wrapping to support ReadByte.

We were converting a non-pointer to an interface,
which forcibly allocates, since interfaces must contain pointers.

Fix that by making the ReadByte methods use pointer receivers.

    name           old time/op    new time/op    delta
    ReadBlocks-16    1.18ms ± 2%    1.15ms ± 5%   -2.73%  (p=0.003 n=11+11)

    name           old speed      new speed      delta
    ReadBlocks-16   441MB/s ± 2%   453MB/s ± 5%   +2.85%  (p=0.003 n=11+11)

    name           old alloc/op   new alloc/op   delta
    ReadBlocks-16    1.33MB ± 0%    1.29MB ± 0%   -2.41%  (p=0.000 n=12+12)

    name           old allocs/op  new allocs/op  delta
    ReadBlocks-16     13.5k ± 0%     11.5k ± 0%  -14.79%  (p=0.000 n=12+12)