Combine API options for simplicity and logical coherence
Introduce a simple `Option` type that includes both read and write
options, as well as read-write options. The rationale for this change is
to avoid re-declaring options that apply to both read and write, and
reduce code verbosity when converting from one type to another as part
of API call chains.

Deprecate `ReadOption`, `WriteOption` and `ReadWriteOption` in favour of
the simpler `Option`, and update APIs to use the new type.

Remove now redundant internal utility functions for working with option
types.

Reformat code using `gofumt` utility for consistency.