add ZeroLengthSectionAsEOF option to the v0 module
This requires plumbing options through NewCarReader and LoadCar.
This change should be backwards compatible with nearly all programs.

The util package is exported, so we can't change its API.
However, when encountering null padding, ReadNode would simply panic,
as LdRead would return an empty []byte and then ReadCid would slice it.

Since those util APIs would panic in that scenario, we can fix that,
and change the behavior to what we need without breaking any users.
In particular, make this edge case return a sentinel error,
so that the user can choose what to do.

In particular, CarReader.Next will convert the error to an io.EOF
if the ZeroLengthSectionAsEOF option is set.
If the option isn't set, we'll simply return the error unchanged,
which is still better than the old panic.

While at it, replace ReadCid with go-cid's ReadFromBytes,
which is better maintained and is not prone to panics.

Finally, this adds the test written by Łukasz,
check that we also do the right thing with the option on/off.

Fixes #171.

Co-Authored-By: Łukasz Magiera <magik6k@gmail.com>