Add API to regenerate index from CARv1 or CARv2
The index generation APIs either allowed reading an existing index from
a CARv2 or explicitly required a CARv1 to generate index.

Introduce APIs to make it easier for users that want to regenerate the
index regardless of whether it exists in a CAR file or not. The index
generation APIs are changed to accept either of the formats and
re-generate the index from the data payload unless `ReadOrGenerate` is
called.

Adjust the tests to run for all flavours of index generation with both
CARv1 and CARv2 payload.