Use a fix code as the multihash code for `CarIndexSorted`
Previous changes added the `ForEach` interface to `Index` type which
enables iteration through the index by multihash and offset. However,
not all index types contain enough information to construct the full
multihash. Namely, `CarIndexSorted` only stores the digest portion of
the multihashes.

In order to implement `ForEach` for this index type correctly uses the
`uint64` max value as the code in the multihash and document the
behaviour where the iterations over this index type should not rely on
the returned code. Note that the max value is used as a code that
doesn't match any existing multicodec.Code to avoid misleading users.