blockstore: allow to pass a file to write in
This allow the caller to control the file lifecycle and implement a cleaning strategy.
An example would be:
- open a file in a temporary folder
- if the write suceed, close the file and move it to its final destination
- on error, remove the file

If the underlying filesystem support it, an anonymous file can be used instead: open
then immediately delete, use the file descriptor to write. A crash before the end of the
write process would release the used storage automatically.